+++ resources/qml_legacy/AlbumsView.qml
@@ -1,7 +1,7 @@
 import QtQuick
 import QtQuick.Controls
 import QtQuick.Layouts
-import SamsungGallery 1.0
+import QGalleryX 1.0
 
 StackView {
     id: stack