+++ src/main.cpp
@@ -57,14 +57,14 @@ int main(int argc, char *argv[]) {
   // Register with AsyncImageProvider (Static)
   AsyncImageProvider::setFrameScheduler(frameScheduler);
 
-  qmlRegisterType<ImageModel>("SamsungGallery", 1, 0, "ImageModel");
+  qmlRegisterType<ImageModel>("QGalleryX", 1, 0, "ImageModel");
   // ...
-  qmlRegisterType<AlbumModel>("SamsungGallery", 1, 0, "AlbumModel");
-  qmlRegisterType<GroupedProxyModel>("SamsungGallery", 1, 0,
+  qmlRegisterType<AlbumModel>("QGalleryX", 1, 0, "AlbumModel");
+  qmlRegisterType<GroupedProxyModel>("QGalleryX", 1, 0,
                                      "GroupedProxyModel");
-  qmlRegisterType<SettingsHelper>("SamsungGallery", 1, 0, "SettingsHelper");
-  qmlRegisterType<SystemMonitor>("SamsungGallery", 1, 0, "SystemMonitor");
-  qmlRegisterType<DesktopHelper>("SamsungGallery", 1, 0, "DesktopHelper");
+  qmlRegisterType<SettingsHelper>("QGalleryX", 1, 0, "SettingsHelper");
+  qmlRegisterType<SystemMonitor>("QGalleryX", 1, 0, "SystemMonitor");
+  qmlRegisterType<DesktopHelper>("QGalleryX", 1, 0, "DesktopHelper");
 
   QQmlApplicationEngine engine;
 
@@ -94,7 +94,7 @@ int main(int argc, char *argv[]) {
       []() { QCoreApplication::exit(-1); }, Qt::QueuedConnection);
 
   // Load Main.qml directly from resources
-  const QUrl url("qrc:/SamsungGallery/resources/qml/Main.qml");
+  const QUrl url("qrc:/QGalleryX/resources/qml/Main.qml");
 
   // Exit if QML fails to load
   QObject::connect(